Evennia updated to release: 5294a19f552d
Added support for character grouping.

Worked on coding a dark room that limits players abilities for use in quests etc.

Added dialogue support to enemy NPC's as well as allowed for all mobs to be set to reanimate.

Added more descriptive text to combat, so no more same boring text over and over.

Created a Parent command class called ActionCommand that allows for prompt generation when needed, but none when its not.

Added some new quests.