hacked menusystem to allow for nodes to call 'self.goto('END') from within node.code, which allows for the last node in a tree to gracefully exit the menu and reset the character model.
added personal lockfuncs for checking if a character is on a quest or has completed a quest.  This is used for quest items view() locks.

Tweaked the first login experience, making it less confusing.

removed a shitty hack to check for quests now that the lockfuncs are written and working.