added lockfuncs to check for skills and spells, removed the nasty hack within the holds() check that was used previously.
character models can now automagically add new attributes to themselves if they detect changes.  This currently only
supports the addition of completely new atttributes, i.e. you cant add a key/value pair to a dict and have the change properly
propagated.

Changed the way the zone runner spawns mobs.  Previously the check to pull all the mobs in to check room mob levels and what not did so by an alias shared by all the mobs in the particular zone.  This works great when there is only ONE instance of said zone, however in the case of dungeons multiple may get the same dungeon template.  This caused all sorts of oddities with mob generation in dungeons, however now the list to check is pulled via an alias based on the mob generator object, which each zone only has one of.

Added some logging to things that needed it.